requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.3.0
requests==2.32.3
beautifulsoup4==4.12.3
gunicorn==22.0.0
